feat: speedup metaGEM env setup env_setup by using mamba instead of conda
Previously was taking ages for conda to solve environment. env_setup.sh will now first create a mamba environment and then use mamba to install the new version of the metaGEM environment with the yml file. The old env file was kept because it has hardcoded version numbers for posterity. The new env file will allow users to download newer versions of tools if they come out in the future.